IG-7075: Common role migrated to repo
The role `common` has been migrated to it's own repo.  The repo is on
our internal github, named roles.junos-common.  The history relating to
the files within the common role was migrated to the new repo as well.

README.md -- Removed information relating to the common role
requirements.yml -- Updated to pull the junos-common role from github
deploy.yml -- Corrected reference to the common role