IG-7819: Clone data repos for deployment added
Previously config files were stored in the `files` directory.  It was
decided to separate the data from the deployment.  The data was moved
into repo's per environment.  The deploy required changes to support
pulling the repo's into the deploy plabyook.  Currently the playbook is
pulling the master branch of  the environment data repo.