IG-7819: Refactor, use with Vagrant & local repo
Vagrant was not accounted for in the previous iteration of
deploy_config.  A new variable static_config_dir is now in place that is
configured within each environment and can be overridden via -e.  When
overriding with -e or using Vagrant, we no longer delete the build_dir
directory nor do we pull down the remote git repository.  To use a local
repo the --skip-tag option will be needed, with remote_repo