IG-7819: `local_config_source` adjusted
Previously the `local_config_source` variable worked for the local
Vagrant environment, but unfortunately did not properly handle local
data repos.  This was discovered while attempting to generate a diff of
changes from the locally checked out data.junos-atl2 repo.  The
appending of the `environment` variable to the end of
`local_config_source` dictated a structure that was incompatible with
local data repos, and only worked with config(s) contained within
./files/